deprecated current aggregation code
