+++ flink-examples/flink-scala-examples/src/main/scala/org/apache/flink/examples/scala/graph/PageRankBasic.scala
@@ -22,7 +22,7 @@ import java.lang.Iterable
 import org.apache.flink.api.common.functions.GroupReduceFunction
 import org.apache.flink.api.scala._
 import org.apache.flink.examples.java.graph.util.PageRankData
-import org.apache.flink.api.java.aggregation.Aggregations.SUM
+import org.apache.flink.api.java.aggregation.deprecated.Aggregations.SUM
 
 import org.apache.flink.util.Collector
