+++ flink-examples/flink-scala-examples/src/main/scala/org/apache/flink/examples/scala/relational/TPCHQuery10.scala
@@ -20,7 +20,7 @@ package org.apache.flink.examples.scala.relational
 import org.apache.flink.api.scala._
 import org.apache.flink.util.Collector
 
-import org.apache.flink.api.java.aggregation.Aggregations
+import org.apache.flink.api.java.aggregation.deprecated.Aggregations
 
 /**
  * This program implements a modified version of the TPC-H query 10.
