+++ flink-java/src/main/java/org/apache/flink/api/java/DataSet.java
@@ -32,7 +32,7 @@ import org.apache.flink.api.common.io.OutputFormat;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.common.operators.base.JoinOperatorBase.JoinHint;
 import org.apache.flink.api.common.operators.base.PartitionOperatorBase.PartitionMethod;
-import org.apache.flink.api.java.aggregation.Aggregations;
+import org.apache.flink.api.java.aggregation.deprecated.Aggregations;
 import org.apache.flink.api.java.functions.FormattingMapper;
 import org.apache.flink.api.java.functions.KeySelector;
 import org.apache.flink.api.java.functions.SelectByMaxFunction;
@@ -269,6 +269,7 @@ public abstract class DataSet<T> {
 	 * @see AggregateOperator
 	 * @see DataSet
 	 */
+	@Deprecated
 	public AggregateOperator<T> aggregate(Aggregations agg, int field) {
 		return new AggregateOperator<T>(this, agg, field, Utils.getCallLocationName());
 	}
@@ -280,6 +281,7 @@ public abstract class DataSet<T> {
 	 *
 	 * @see org.apache.flink.api.java.operators.AggregateOperator
 	 */
+	@Deprecated
 	public AggregateOperator<T> sum (int field) {
 		return this.aggregate (Aggregations.SUM, field);
 	}
@@ -297,8 +299,9 @@ public abstract class DataSet<T> {
 	 * @see #aggregate(Aggregations, int)
 	 * @see #maxBy(int...)
 	 */
-	public AggregateOperator<T> max(int field) {
-		return aggregate(Aggregations.MAX, field);
+	@Deprecated
+	public AggregateOperator<T> max (int field) {
+		return aggregate (Aggregations.MAX, field);
 	}
 
 	/**
@@ -314,8 +317,9 @@ public abstract class DataSet<T> {
 	 * @see #aggregate(Aggregations, int)
 	 * @see #minBy(int...)
 	 */
-	public AggregateOperator<T> min(int field) {
-		return aggregate(Aggregations.MIN, field);
+	@Deprecated
+	public AggregateOperator<T> min (int field) {
+		return aggregate (Aggregations.MIN, field);
 	}
 	
 	/**
