+++ flink-java/src/main/java/org/apache/flink/api/java/aggregation/deprecated/AvgAggregationFunction.java
@@ -1,4 +1,4 @@
-package org.apache.flink.api.java.aggregation;
+package org.apache.flink.api.java.aggregation.deprecated;
 /**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
