+++ flink-java/src/test/java/org/apache/flink/api/java/operator/AggregateOperatorTest.java
@@ -24,8 +24,8 @@ import java.util.List;
 import org.junit.Assert;
 import org.apache.flink.api.common.InvalidProgramException;
 import org.apache.flink.api.common.typeinfo.BasicTypeInfo;
-import org.apache.flink.api.java.aggregation.Aggregations;
-import org.apache.flink.api.java.aggregation.UnsupportedAggregationTypeException;
+import org.apache.flink.api.java.aggregation.deprecated.Aggregations;
+import org.apache.flink.api.java.aggregation.deprecated.UnsupportedAggregationTypeException;
 import org.apache.flink.api.java.tuple.Tuple5;
 import org.apache.flink.api.java.typeutils.TupleTypeInfo;
 import org.junit.Test;
