+++ flink-scala/src/main/scala/org/apache/flink/api/scala/AggregateDataSet.scala
@@ -17,7 +17,7 @@
  */
 package org.apache.flink.api.scala
 
-import org.apache.flink.api.java.aggregation.Aggregations
+import org.apache.flink.api.java.aggregation.deprecated.Aggregations
 import org.apache.flink.api.scala.operators.ScalaAggregateOperator
 
 import scala.reflect.ClassTag
